Reforço de segurança

+++ src/main/java/br/gov/es/participe/controller/CitizenController.java
@@ -5,6 +5,7 @@ import br.gov.es.participe.controller.dto.PersonKeepCitizenDto;
 import br.gov.es.participe.controller.dto.PersonParamDto;
 import br.gov.es.participe.service.LocalityService;
 import br.gov.es.participe.service.PersonService;
+import br.gov.es.participe.util.dto.MessageDto;
 import br.gov.es.participe.util.interfaces.ApiPageable;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -30,6 +31,7 @@ public class CitizenController {
   @ApiPageable
   @GetMapping
   public ResponseEntity<Page<PersonKeepCitizenDto>> listCitizen(
+      @RequestHeader("Authorization") String token,
       @RequestParam(value = "name", required = false, defaultValue = "") String name,
       @RequestParam(value = "email", required = false, defaultValue = "") String email,
       @RequestParam(value = "autentication", required = false, defaultValue = "") String autentication,
@@ -37,44 +39,84 @@ public class CitizenController {
       @RequestParam(value = "locality", required = false, defaultValue = "") List<Long> locality,
       @RequestParam(value = "conferenceId", required = false) Long conferenceId,
       @ApiIgnore Pageable page) {
-    Page<PersonKeepCitizenDto> list = personService
-        .listKeepCitizen(name, email, autentication, active, locality, conferenceId, page);
-    return ResponseEntity.status(200).body(list);
+
+    if (personService.hasOneOfTheRoles(token, new String[] { "Administrator" })) {
+      Page<PersonKeepCitizenDto> list = personService
+          .listKeepCitizen(name, email, autentication, active, locality, conferenceId, page);
+      return ResponseEntity.status(200).body(list);
+    } else {
+      return ResponseEntity.status(401).body(null);
+    }
+
   }
 
   @GetMapping("/{personId}")
-  public ResponseEntity<PersonKeepCitizenDto> getCitizenById(@PathVariable Long personId,
+  public ResponseEntity<PersonKeepCitizenDto> getCitizenById(
+      @RequestHeader("Authorization") String token,
+      @PathVariable Long personId,
       @RequestParam Long conferenceId) {
-    PersonKeepCitizenDto citizen = personService.findCitizenById(personId, conferenceId);
-    return ResponseEntity.status(200).body(citizen);
+
+    if (personService.hasOneOfTheRoles(token, new String[] { "Administrator" })) {
+      PersonKeepCitizenDto citizen = personService.findCitizenById(personId, conferenceId);
+      return ResponseEntity.status(200).body(citizen);
+    } else {
+      return ResponseEntity.status(401).body(null);
+    }
   }
 
   @SuppressWarnings({ "rawtypes" })
   @PostMapping
-  public ResponseEntity store(@RequestBody PersonParamDto personParam) {
-    return personService.storePerson(personParam, false);
+  public ResponseEntity store(
+      @RequestHeader("Authorization") String token,
+      @RequestBody PersonParamDto personParam) {
+
+    if (personService.hasOneOfTheRoles(token, new String[] { "Administrator" })) {
+      return personService.storePerson(personParam, false);
+    } else {
+      return ResponseEntity.status(401).body(null);
+    }
   }
 
   @SuppressWarnings({ "rawtypes" })
   @DeleteMapping("/{id}")
-  public ResponseEntity destroy(@PathVariable Long id) {
-    personService.delete(id);
-    return ResponseEntity.status(200).build();
+  public ResponseEntity destroy(
+      @RequestHeader(name = "Authorization") String token,
+      @PathVariable Long id) {
+
+    if (personService.hasOneOfTheRoles(token, new String[] { "Administrator" })) {
+      personService.delete(id);
+      return ResponseEntity.status(200).build();
+    } else {
+      return ResponseEntity.status(401).body(null);
+    }
   }
 
   @SuppressWarnings({ "rawtypes" })
   @PutMapping("/{id}")
-  public ResponseEntity update(@RequestHeader(name = "Authorization") String token, @PathVariable Long id,
+  public ResponseEntity update(
+      @RequestHeader(name = "Authorization") String token,
+      @PathVariable Long id,
       @RequestBody PersonParamDto personParam) {
-    personParam.setId(id);
-    return personService.updatePerson(personParam, false);
+
+    if (personService.hasOneOfTheRoles(token, new String[] { "Administrator" })) {
+      personParam.setId(id);
+      return personService.updatePerson(personParam, false);
+    } else {
+      return ResponseEntity.status(401).body(null);
+    }
   }
 
   @GetMapping("/localities/{idConferences}")
   public ResponseEntity<LocalityCitizenSelectDto> getLocalities(
+      @RequestHeader(name = "Authorization") String token,
       @PathVariable(name = "idConferences", required = false) Long idConferences,
       @RequestParam(name = "name", required = false, defaultValue = "") String name) {
-    LocalityCitizenSelectDto localityCitizenSelectDto = localityService.getLocalitiesToDisplay(idConferences, name);
-    return ResponseEntity.status(200).body(localityCitizenSelectDto);
+
+    if (personService.hasOneOfTheRoles(token, new String[] { "Administrator" })) {
+      LocalityCitizenSelectDto localityCitizenSelectDto = localityService.getLocalitiesToDisplay(idConferences, name);
+      return ResponseEntity.status(200).body(localityCitizenSelectDto);
+    } else {
+      return ResponseEntity.status(401).body(null);
+    }
   }
 }
